enhance: custom node multithreading
Extend the node with better handling of the subscription to the image topic. If process_image() will take more time to run it will also block the execution of rclpy.spin_once(self, timeout_sec=0.05) that is needed to trigger the image_callback(). So let's move the spin functionality to a separate thread