fix: relay camera_info to solve rviz issue
It's because RViz always expect the image and the camera_info topics with the same prefix which works well for:

/camera/image → /camera/camera_info

But doesn't work for:

/camera/image/compressed → /camera/image/camera_info

because we don't publish the camera_info to that topic. We could remap the camera_info to that topic, but then the uncompressed image won't work in RViz, so it's not the desired solution.